Migrate TextDataset to 2.0 Dataset
