+++ tensorflow_io/text/python/ops/text_ops.py
@@ -18,13 +18,11 @@ from __future__ import division
 from __future__ import print_function
 
 import tensorflow as tf
-from tensorflow import dtypes
-from tensorflow.compat.v1 import data
+from tensorflow_io.core.python.ops import data_ops as data_ops
 from tensorflow_io.core.python.ops import core_ops as text_ops
 
-class TextDataset(data.Dataset):
-  """A Text Dataset
-  """
+class TextDataset(data_ops.Dataset):
+  """A Text Dataset"""
 
   def __init__(self, filename, batch=None):
     """Create a Text Reader.
@@ -32,33 +30,16 @@ class TextDataset(data.Dataset):
     Args:
       filename: A `tf.string` tensor containing one or more filenames.
     """
-    self._data_input = text_ops.text_input(filename, ["none", "gz"])
-    self._batch = 0 if batch is None else batch
-    super(TextDataset, self).__init__()
+    batch = 0 if batch is None else batch
+    dtypes = [tf.string]
+    shapes = [
+        tf.TensorShape([])] if batch == 0 else [
+            tf.TensorShape([None])]
+    super(TextDataset, self).__init__(
+        text_ops.text_dataset,
+        text_ops.text_input(filename, ["none", "gz"]),
+        batch, dtypes, shapes)
 
-  def _inputs(self):
-    return []
-
-  def _as_variant_tensor(self):
-    return text_ops.text_dataset(
-        self._data_input,
-        self._batch,
-        output_types=self.output_types,
-        output_shapes=self.output_shapes)
-
-  @property
-  def output_shapes(self):
-    return tuple([
-        tf.TensorShape([])]) if self._batch == 0 else tuple([
-            tf.TensorShape([None])])
-
-  @property
-  def output_classes(self):
-    return tf.Tensor
-
-  @property
-  def output_types(self):
-    return tuple([dtypes.string])
 
 class TextOutputSequence(object):
   """TextOutputSequence"""
